creates separate folder for nav component
